# asset-service/docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Asset Management Service API
  description: |
    Asset Management Service API
    
    Comprehensive asset lifecycle management including:
    - Asset CRUD operations
    - Master data management (Categories, Makes, Models, Vendors, etc.)
    - User-Asset linking
    - Compliance and validation
    - Document management
    - Warranty and AMC management
    - Audit logging
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8082
    description: Local development server
  - url: https://api.example.com/asset-service
    description: Production server
...
paths:
...
  /api/asset/v1/vendors:
    get:
      tags:
        - Vendors
      summary: List all vendors
      responses:
        '200':
          description: List of vendors

    post:
      tags:
        - Vendors
      summary: Create vendor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorRequest'
      responses:
        '200':
          description: Vendor created
  /api/asset/v1/vendors/{id}:
    get:
      tags:
        - Vendors
      summary: Get vendor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Vendor found

    put:
      tags:
        - Vendors
      summary: Update vendor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorRequest'
      responses:
        '200':
          description: Vendor updated

    delete:
      tags:
        - Vendors
      summary: Delete vendor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Vendor deleted
  /api/asset/v1/vendors/bulk:
    post:
      tags:
        - Vendors
      summary: Bulk create vendors
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/VendorRequest'
      responses:
        '200':
          description: Bulk creation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkUploadResponse'
...
  /api/asset/v1/userlinks/link:
    post:
      tags:
        - User Links
      summary: Link entity to user
      description: Link an asset, component, or other entity to a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetUserUniversalLinkRequest'
      responses:
        '200':
          description: Entity linked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapper'
  /api/asset/v1/userlinks/delink:
    post:
      tags:
        - User Links
      summary: Delink entity from user
      description: Remove the link between an entity and a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetUserUniversalLinkRequest'
      responses:
        '200':
          description: Entity delinked successfully
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ResponseWrapper:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
...
    VendorRequest:
      type: object
      required:
        - vendorName
      properties:
        vendorName:
          type: string
        contactPerson:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
...
    BulkUploadResponse:
      type: object
      properties:
        total:
          type: integer
        success:
          type: integer
        failed:
          type: integer
        results:
          type: array
          items:
            type: object
            properties:
              success:
                type: boolean
              message:
                type: string
              data:
                type: object
    AssetUserUniversalLinkRequest:
      type: object
      required:
        - entityType
        - entityId
        - targetUserId
        - targetUsername
      properties:
        entityType:
          type: string
          enum: [ASSET, COMPONENT, MODEL, MAKE, AMC, WARRANTY, DOCUMENT]
        entityId:
          type: integer
          format: int64
        targetUserId:
          type: integer
          format: int64
        targetUsername:
          type: string
        userId:
          type: integer
          format: int64
        username:
          type: string
...